try fixing issue with user not getting added to database

+++ game_checker.py
@@ -150,14 +150,14 @@ class PSTelegramUsers(db.Model):
 
 
 class XboxTelegramUsers(db.Model):
-    chatID = db.Column(db.Integer, primary_key=True)
+    chatID = db.Column(db.BigInteger, primary_key=True)
     subscribed = db.Column(db.Boolean, nullable=False)
     unsubscribed_games = db.Column(MutableList.as_mutable(db.PickleType), default=[])
     subscribed_games = db.Column(MutableList.as_mutable(db.PickleType), default=[])
 
 
 class SwitchTelegramUsers(db.Model):
-    chatID = db.Column(db.Integer, primary_key=True)
+    chatID = db.Column(db.BigInteger, primary_key=True)
     subscribed = db.Column(db.Boolean, nullable=False)
     unsubscribed_games = db.Column(MutableList.as_mutable(db.PickleType), default=[])
     subscribed_games = db.Column(MutableList.as_mutable(db.PickleType), default=[])
